encoders/generic_trainable/AutoKerasEncoder/requirements.txt
tensorflow-cpu==2.4.0
autokeras==1.0.12
